requirements.txt
# LangChain core
langchain
langchain-core

# OpenAI integration
langchain-openai
openai

# Anthropic integration
langchain-anthropic

# Google Gemini integration
langchain-google-genai
google-generativeai

# Hugging Face integration
langchain-huggingface
transformers
huggingface-hub

#Groq integration
groq
langchain-groq

# Environment variable management
python-dotenv

# Machine learning utilities
numpy
scikit-learn
